Camera Cover Status Checked
This update now checks the camera cover to see if it is open.  If not,
Lanterne will not turn the LEDs on.  (This behavior can be disabled from
the new Preferences form available from the drop-down menu.)  Also, a
continuous-pulse torch mode added to main window, and a continuous loop
option added to Morse code window.